+++ mass_circular_weighing/__init__.py
@@ -1,8 +1,8 @@
 from .log import log
 
 __author__ = 'Measurement Standards Laboratory of New Zealand'
-__copyright__ = '\xa9 2022, ' + __author__
-__version__ = '1.0.8.dev0'
+__copyright__ = '\xa9 2023, ' + __author__
+__version__ = '1.0.10.dev0'
 
 from .gui.gui import show_gui
 from .utils.poll_omega_logger import poll_omega_logger