+++ mass_circular_weighing/routines/run_circ_weigh.py
@@ -41,7 +41,7 @@ def check_for_existing_weighdata(folder, url, se):
             os.makedirs(back_up_folder)
         new_index = len(os.listdir(back_up_folder))  # counts number of files in backup folder
         new_file = os.path.join(back_up_folder, se + '_backup{}.json'.format(new_index))
-        existing_root.is_read_only = False
+        existing_root.read_only = False
         log.debug('Existing root is '+repr(existing_root))
         root = JSONWriter()
         root.set_root(existing_root)
@@ -121,10 +121,9 @@ def do_circ_weighing(bal, se, root, url, run_id, callback1=None, callback2=None,
     -------
     msl.io root object if weighing was completed, False if weighing was not started, or None if weighing was aborted.
     """
-    local_backup_file = os.path.join(local_backup_folder, url.split('\\')[-1])
-
     metadata['Program Version'] = __version__
-    metadata['Mmt Timestamp'] = datetime.now().strftime('%d-%m-%Y %H:%M')
+    timestamp = datetime.now()
+    metadata['Mmt Timestamp'] = timestamp.strftime('%d-%m-%Y %H:%M:%S')
     metadata['Time unit'] = 'min'
     metadata['Ambient monitoring'] = bal.ambient_details
     metadata['Weighing complete'] = False
@@ -146,7 +145,9 @@ def do_circ_weighing(bal, se, root, url, run_id, callback1=None, callback2=None,
              '\nWeight groups are positioned as follows:' +
              '\n' + positionstr.strip('\n'))
 
-    ambient_pre = check_ambient_pre(bal.ambient_instance, bal.ambient_details)
+    log.info(f"Weighing starting at {metadata['Mmt Timestamp']}")
+
+    ambient_pre = check_ambient_pre(bal.ambient_instance, bal.ambient_details, bal.mode)
     if not ambient_pre:
         log.info('Measurement not started due to unsuitable ambient conditions')
         return False
@@ -180,40 +181,58 @@ def do_circ_weighing(bal, se, root, url, run_id, callback1=None, callback2=None,
                 times.append(time)
                 weighdata[cycle, i, :] = [time, reading]
                 if reading is not None:
-                    try:
-                        root.save(file=url, mode='w', ensure_ascii=False)
-                    except OSError:
-                        root.save(file=local_backup_file, mode='w', ensure_ascii=False)
-                        log.warning('Data saved to local backup at '+local_backup_file)
+                    network_ok = save_data(root, url, local_backup_folder, run_id, timestamp)
+                    if not network_ok:
+                        metadata['Network issues'] = True
                 bal.unload_bal(mass, positions[i])
         break
 
     while not bal.want_abort:
-        ambient_post = check_ambient_post(ambient_pre, bal.ambient_instance, bal.ambient_details)
+        ambient_post = check_ambient_post(ambient_pre, bal.ambient_instance, bal.ambient_details, bal.mode)
         for key, value in ambient_post.items():
             metadata[key] = value
 
         metadata['Weighing complete'] = True
+        end_time = datetime.now()
+        metadata['Mmt end time'] = end_time.strftime('%d-%m-%Y %H:%M:%S')
+        log.info(f"Weighing completed at {metadata['Mmt end time']}")
+        elapsed_duration(end_time - timestamp)  # reports weighing duration to log window
+
         weighdata.add_metadata(**metadata)
-        try:
-            root.save(file=url, mode='w', ensure_ascii=False)
-        except:
+        ok = save_data(root, url, local_backup_folder, run_id, timestamp)
+        if not ok:
             log.debug('weighdata:\n' + str(weighdata[:, :, :]))
-            root.save(file=local_backup_file, mode='w', ensure_ascii=False)
-            log.warning('Data saved to local backup: ' + local_backup_file)
 
         return root
 
     log.info('Circular weighing sequence aborted')
     if reading:
         weighdata.add_metadata(**metadata)
-        try:
-            root.save(file=url, mode='w', ensure_ascii=False)
-        except OSError:
-            root.save(file=local_backup_file, mode='w', ensure_ascii=False)
-            log.warning('Data saved to local backup file: ' + local_backup_file)
+        ok = save_data(root, url, local_backup_folder, run_id, timestamp)
+        if not ok:
+            log.debug('weighdata:\n' + str(weighdata[:, :, :]))
 
     return None
+
+
+def save_data(root, url, local_backup_folder, run_id, timestamp):
+    """Saves data to local drive and attempts to also save to network drive"""
+    local_folder = os.path.join(local_backup_folder, os.path.split(os.path.dirname(url))[-1])
+    # ensure a unique filename in case of intermittent internet
+    local_file = os.path.join(
+        local_folder,
+        os.path.basename(url).strip('.json') + f'_{run_id}_{timestamp.strftime("%Y%m%d_%H%M%S")}.json'
+    )
+    if not os.path.exists(local_folder):
+        os.makedirs(local_folder)
+    root.save(file=local_file, mode='w', encoding='utf-8', ensure_ascii=False)
+    try:
+        root.save(file=url, mode='w', encoding='utf-8', ensure_ascii=False)
+        return True
+    except FileNotFoundError:
+        log.warning(f'Unable to save to {url}. Please check network connection.')
+        log.info(f"Data saved to {local_file}")
+        return False
 
 
 def analyse_weighing(root, url, se, run_id, bal_mode, timed=False, drift=None, EXCL=3, local_backup_folder=local_backup, **metadata):
@@ -250,7 +269,7 @@ def analyse_weighing(root, url, se, run_id, bal_mode, timed=False, drift=None, E
     if not weighdata.metadata.get('Weighing complete'):
         return None
 
-    log.info('CIRCULAR WEIGHING ANALYSIS for scheme entry '+ se + ' ' + run_id)
+    log.info(f'CIRCULAR WEIGHING ANALYSIS for scheme entry {se} {run_id}')
 
     weighing = CircWeigh(se)
     if timed:
@@ -289,7 +308,7 @@ def analyse_weighing(root, url, se, run_id, bal_mode, timed=False, drift=None, E
 
     analysis_meta = {
         'Program Version': __version__,
-        'Analysis Timestamp': datetime.now().strftime('%d-%m-%Y %H:%M'),
+        'Analysis Timestamp': datetime.now().strftime('%d-%m-%Y %H:%M:%S'),
         'Residual std devs': str(weighing.stdev),
         'Selected drift': drift,
         'Uses mmt times': timed,
@@ -313,17 +332,16 @@ def analyse_weighing(root, url, se, run_id, bal_mode, timed=False, drift=None, E
 
     flag = weighdata.metadata.get('Ambient OK?')
     if not flag:
-        log.warning('Analysis not accepted due to ambient conditions during weighing')
-        analysis_meta['Acceptance met?'] = False
+        if flag is False:
+            log.warning('Analysis not accepted due to unsuitable ambient conditions during weighing')
+            analysis_meta['Acceptance met?'] = False
+        else:  # e.g. flag is None
+            log.warning('Ambient conditions unavailable during weighing')
 
     weighanalysis.add_metadata(**analysis_meta)
 
-    try:
-        root.save(file=url, mode='w', encoding='utf-8', ensure_ascii=False)
-    except OSError:
-        local_backup_file = os.path.join(local_backup_folder, url.split('\\')[-1])
-        root.save(file=local_backup_file, mode='w', ensure_ascii=False)
-        log.warning('Data saved to local backup file: ' + local_backup_file)
+    timestamp = datetime.strptime(weighdata.metadata.get('Mmt Timestamp'), '%d-%m-%Y %H:%M:%S')
+    save_data(root, url, local_backup_folder, run_id, timestamp)  # save to same file on C: drive as the weighing data
 
     log.info('Circular weighing analysis for '+se+', '+run_id+' complete\n')
 
@@ -419,3 +437,13 @@ def check_existing_runs(root, scheme_entry, display_message=False):
     run_1_no = int(run_id.strip('run_'))
 
     return good_runs, run_1_no  # returns integers for each
+
+
+def elapsed_duration(duration):
+    duration_in_s = duration.total_seconds()
+    hours = int(divmod(duration_in_s, 3600)[0])  # Seconds in an hour = 3600
+    minutes = int(divmod(duration_in_s, 60)[0])  # Seconds in a minute = 60
+    seconds = int(divmod(duration_in_s, 60)[1])
+    log.info(f"(duration of {hours} hours, {minutes} minutes and {seconds} seconds)")
+
+    return hours, minutes, seconds